requirements.txt
Django
djangorestframework
gunicorn
psycopg2-binary
whitenoise